refactor(core): rename admit_loop → admit
The migration is done; the "_loop" suffix was a parallel-path marker.
Rename the module to its final name and update every import + docstring
reference across src, tests, and packs. No behavior change.
