+++ packs/cyber_webapp/cyber_webapp/families/pentest.py
@@ -37,7 +37,7 @@ from openrange.core.contracts import (
 from openrange.world_ir import Node, Visibility, WorldGraph
 
 if TYPE_CHECKING:
-    from openrange.core.admit_loop import Snapshot
+    from openrange.core.admit import Snapshot
 
 
 class WebappPentest(TaskFamily):
