+++ tests/test_cyber_auto_curriculum.py
@@ -29,7 +29,7 @@ from cyber_webapp import WebappPack, WebappPentest
 from cyber_webapp.mutation import available_mutations
 from cyber_webapp.vulnerabilities import CATALOG as VULN_CATALOG
 
-from openrange.core.admit_loop import Snapshot
+from openrange.core.admit import Snapshot
 from openrange.core.contracts import EpisodeReportLike, Mutation
 from openrange.world_ir import GraphPatch, WorldGraph, apply_patch
